fix: Browse file button

+++ backend/src/main/java/com/dat_management/backend/dto/SystemConfigRequest.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class SystemConfigRequest {
 
-    private Integer fileUploadSizeMb;
+    private Double fileUploadSizeMb;
     private Integer sessionTimeoutMinutes;
     private Integer jwtExpiryHours;
     private Integer maxLoginAttempts;
